WM-1414 Refactoring WesRunLog to omit Cromwell's "workflowLog" object (#6925)
