+++ engine/src/main/scala/cromwell/webservice/routes/wes/WesRunLog.scala
@@ -1,6 +1,5 @@
 package cromwell.webservice.routes.wes
 
-import cromwell.webservice.routes.wes.WesState.WesState
 import spray.json.JsObject
 
 
@@ -20,15 +19,3 @@ final case class WesRunRequest(workflow_params: Option[JsObject],
                                workflow_engine_parameters: Option[JsObject],
                                workflow_url: Option[String]
                               )
-
-final case class WesRunLog(run_id: String,
-                           request: WesRunRequest,
-                           state: WesState,
-                           run_log: Option[WesLog],
-                           task_logs: Option[List[WesLog]],
-                           outputs: Option[JsObject]
-                          )
-
-object WesRunLog {
-  def fromJson(json: String): WesRunLog = CromwellMetadata.fromJson(json).wesRunLog
-}
